+++ app/src/main/java/ru/dvn/gitapp/di/RemoteModule.kt
@@ -1,0 +1,50 @@
+package ru.dvn.gitapp.di
+
+import dagger.Module
+import dagger.Provides
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.adapter.rxjava3.RxJava3CallAdapterFactory
+import retrofit2.converter.gson.GsonConverterFactory
+import ru.dvn.gitapp.data.remote.RemoteUsersRepositoryImpl
+import ru.dvn.gitapp.data.remote.user.GithubApi
+import javax.inject.Singleton
+
+@Module
+class RemoteModule {
+    private val baseUrl = "https://api.github.com/"
+
+    @Provides
+    @Singleton
+    fun provideRetrofit(): Retrofit {
+        return Retrofit.Builder()
+            .baseUrl(baseUrl)
+            .addConverterFactory(GsonConverterFactory.create())
+            .addCallAdapterFactory(RxJava3CallAdapterFactory.create())
+            .client(buildHttpClient())
+            .build()
+    }
+
+    @Provides
+    @Singleton
+    fun provideGithubApi(retrofit: Retrofit): GithubApi {
+        return retrofit.create(GithubApi::class.java)
+    }
+
+    @Provides
+    @Singleton
+    fun provideRemoteRepository(api: GithubApi): RemoteUsersRepositoryImpl {
+        return RemoteUsersRepositoryImpl(api)
+    }
+
+    private fun buildHttpClient(): OkHttpClient {
+        return OkHttpClient.Builder().apply {
+            addInterceptor(
+                HttpLoggingInterceptor().apply {
+                    level = HttpLoggingInterceptor.Level.BODY
+                }
+            )
+        }.build()
+    }
+}
